moved service from features to general service folder
